Cleaned up PHPDoc to resolve IDE warnings.
There were a few instances of missing, incorrect, or inconsistent documentation in the PHPDoc docblocks, which were causing PHPStorm to complain about accessing missing properties and methods, or simply not provide any hinting when working with Objects from the library. Barring some remaining complains about unused namespaces in the XML string literals, these should all be fixed now.

This is purely a tooling improvement. It doesn't modify the functionality or the API in any way.