Added ability to access public spreadsheets.
These are spreadsheets that have been published using the "Publish to the web" feature (not to be confused with those that have been made public using the "Public on the web" visibility option).

The ServiceRequest should be initialised with the sheet's key as its $accessToken, and $tokenType should be set to "basic".  You can then use the PublicSpreadsheetService to access the sheet.

Example:

```php
use Google\Spreadsheet\DefaultServiceRequest;
use Google\Spreadsheet\ServiceRequestFactory;

$serviceRequest = new DefaultServiceRequest($key, "basic");
ServiceRequestFactory::setInstance($serviceRequest);
$spreadsheetService = new Google\Spreadsheet\PublicSpreadsheetService();
```

I have exposed the $accessToken in the DefaultServiceRequest class, via a new getAccessToken() method, so that this can be accessed from the PublicSpreadsheetService class.  The new PublicSpreadsheetService class is based on SpreadsheetService, but does not implement the getSpreadsheetById() method, as this is not possible and makes no sense in the context of a public sheet.  There seemed little point extending the basic class, as the code is not sufficiently re-usable between the two contexts.